pyproject.toml
[project]
name = "chess-ai-tutor"
version = "0.1.0"
description = "Pedagogical chess analysis system with Stockfish + fine-tuned LLM"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Zheng"}
]

dependencies = [
    "python-chess>=1.10",
    "mcp>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "cairosvg>=2.7",
    "pillow>=10.0",
    "click>=8.1",
    "rich>=13.0",
    "openai>=2.15.0",
    "fastapi[standard]>=0.115",
    "datasets>=2.18",
    "flash-attn-3>=3.0.0",
    "trl==0.29.1",
    "jmespath>=1.0",
]

[project.optional-dependencies]
training = [
    "transformers>=5.0",
    "huggingface_hub>=1.3.0",
    "datasets>=2.18",
    "accelerate>=0.30",
    "bitsandbytes>=0.43",
    "peft>=0.10",
    "liger-kernel>=0.7",
    "wandb",
    "deepspeed",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
chess-mcp = "chess_mcp.server:main"
chess-tutor = "tutor.cli:main"
chess-review = "tutor.review:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chess_mcp", "src/verification", "src/tutor"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
]
